backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
pymongo==4.11.3
openai==1.68.2
python-dotenv==1.1.0
PyJWT==2.8.0
